Move the functions that have only single use to the private package
Some validation functions are used only from `Validate` on `Revision`
objects and there's no need to keep them in public space, thus
increasing the API surface, number of exported objects, etc.

So move them to the revision implementation file along with the tests.
